revert "fiddle with data types"
it's stupid, it may actually be hurting performance instead even if it
isn't doing some heavy calculations···

+++ shoot.h
@@ -21,7 +21,7 @@ typedef enum trans_t {
 extern int key;
 
 // level tracking
-extern uint8_t levels;
+extern unsigned int levels;
 extern uintptr_t level;
 
 // general engine functions
